==> 03_RandomForest_scaling.ipynb <==
## Explore how the size of a grid affects the speed and accuracy of Random Forests

First create a Test and Train set for evaluation

import pandas as pd
data_complete = pd.read_csv("data/train.csv")

data_complete.describe()


data_complete.dtypes

from sklearn.cross_validation import train_test_split

data_complete.place_id = data_complete.place_id.astype(str)
data_complete.dtypes


data_complete.row_id.head()

features = data_complete[['x', 'y', 'accuracy', 'time']]
labels = data_complete [['place_id']]

features_train, features_test, labels_train, labels_test = train_test_split(
    features, labels, test_size=0.33)

features_train.head()

labels_train.head()

labels_train.count()

labels_test.count()

#A function to take a x and y range and then build and test a random forest
import time
from sklearn.ensemble import RandomForestClassifier
from sklearn import preprocessing

def mlExplore(x_range, y_range):
    #Filter the data so that the bottom left of the grid is at 5,5
    x_filter_train = (features_train.x < 5 + x_range) & (features_train.x > 5)
    y_filter_train = (features_train.y < 5 + y_range) & (features_train.y > 5)
    filter_train = x_filter_train & y_filter_train
    
    x_filter_test = (features_test.x < 5 + x_range) & (features_test.x > 5)
    y_filter_test = (features_test.y < 5 + y_range) & (features_test.y > 5)
    filter_test = x_filter_test & y_filter_test
    
    grid_features_train = features_train[filter_train]
    grid_labels_train = labels_train[filter_train]
    
    grid_features_test = features_test[filter_test]
    grid_labels_test = labels_test[filter_test]
    
    #Record how long the machine learning takes
    time_start = time.time()
    
    #Convert the labels to an array binary array [n_obs, n_unique_labels]
    lb = preprocessing.LabelBinarizer()
    all_possible_labels= pd.concat([grid_labels_test, grid_labels_train])
    lb_fitted = lb.fit(all_possible_labels)
    grid_labels_test = lb_fitted.transform(grid_labels_test)
    grid_labels_train = lb_fitted.transform(grid_labels_train)
    
    #Count the number of points
    num_training_obs, num_labels = grid_labels_train.shape
    
    
    
    rf = RandomForestClassifier() # defualt random forest
    fitted = rf.fit(grid_features_train, grid_labels_train)
    score = fitted.score(grid_features_test, grid_labels_test)
    
    time_end = time.time()
    seconds_elapsed = time_end - time_start
    

    
    return(num_training_obs, num_labels, seconds_elapsed, score)
    

mlExplore(0.5, 0.01) # (x_range, y_range)

mlExplore(0.5, 0.05)


mlExplore(1, 0.1)



mlExplore(0.2, 0.02)

mlExplore(0.1, 0.01)

mlExplore(2, 0.1) # Range is about 3 std each way

import numpy as np

for x_range in np.arange(0.1, 0.7, 0.1):
    print(x_range, mlExplore(x_range, 0.03))

for y_range in np.arange(0.01, 0.05, 0.01):
    print(y_range, mlExplore(0.4, y_range))

mlExplore(1, 0.05)

mlExplore(1.5, 0.06)

mlExplore(0.75, 0.025)

mlExplore(1, 0.04)

mlExplore(0.75, 0.03)

Explore in more robust way
Exploratory analysis showed xspread to be 25 times the y spread so we will fix at this ration
Loop over a range of values and plot to see where score starts to break down

import numpy as np
xs = [0.75, 0.5, 0.25, 0.1, 0.075, 0.05, 0.04, 0.03]
results = {"range": xs, "obs": [], "places": [], "time": [], "score": []}
for x_range in xs:
    y_range = x_range / 25.0
    a,b,c,d = mlExplore(x_range, y_range)
    results["obs"].append(a)
    results["places"].append(b)
    results["time"].append(round(c,2))
    results["score"].append(round(d,2))

results

import numpy as np
xs = [0.5,0.3, 0.2, 0.1, 0.09, 0.08, 0.07,0.06, 0.055, 0.05, 0.045]
results = {"range": xs, "obs": [], "places": [], "time": [], "score": []}
for x_range in xs:
    y_range = x_range / 25.0
    a,b,c,d = mlExplore(x_range, y_range)
    results["obs"].append(a)
    results["places"].append(b)
    results["time"].append(round(c,2))
    results["score"].append(round(float(d),2))
pd.DataFrame(results)

Consider using lots of small square of range 0.075. Remember that the data in the table above will be 30% larger as we split into train and test

 
 Possibly create the actual evaluation metric too
 
 Possibly make the grid 3d - with accuracy as the 3rd dimension - again set the ratio of the sides of the grid equal to the ration of the features standard deviation for a particular place *Accuracy is not linearly distributed*
 
 If each grid is 0.075*0.003 then the number of grids is (10/0.075)*2 x (10/0.003)*2 = 1.8Million (Still less than the number of points? Test set has 8 Million.
 
 To execute in spark we will round the x and ys to 0.075 and 0.003 and then group the data to get 1.8Million points and associated string of test_row_ids
 
 Building 2 Million Random Forest models might be slow?

## Repeat above but time the whole function (including filtering data)

1e6 / 60/60

#A function to take a x and y range and then build and test a random forest
import time
from sklearn.ensemble import RandomForestClassifier
from sklearn import preprocessing

def mlExplore(x_range, y_range):
    #Filter the data so that the bottom left of the grid is at 5,5
    x_filter_test = (features_test.x < 5 + x_range) & (features_test.x > 5)
    y_filter_test = (features_test.y < 5 + y_range) & (features_test.y > 5)
    filter_test = x_filter_test & y_filter_test
    
    grid_features_test = features_test[filter_test]
    grid_labels_test = labels_test[filter_test]
    
    #Record how long the machine learning takes and filtering takes
    time_start = time.time()
    x_filter_train = (features_train.x < 5 + x_range) & (features_train.x > 5)
    y_filter_train = (features_train.y < 5 + y_range) & (features_train.y > 5)
    filter_train = x_filter_train & y_filter_train
    
    grid_features_train = features_train[filter_train]
    grid_labels_train = labels_train[filter_train]
    
   
    
    #Convert the labels to an array binary array [n_obs, n_unique_labels]
    lb = preprocessing.LabelBinarizer()
    all_possible_labels= pd.concat([grid_labels_test, grid_labels_train])
    lb_fitted = lb.fit(all_possible_labels)
    grid_labels_test = lb_fitted.transform(grid_labels_test)
    grid_labels_train = lb_fitted.transform(grid_labels_train)
    
    #Count the number of points
    num_training_obs, num_labels = grid_labels_train.shape
    
    
    
    rf = RandomForestClassifier() # defualt random forest
    fitted = rf.fit(grid_features_train, grid_labels_train)
    score = fitted.score(grid_features_test, grid_labels_test)
    
    time_end = time.time()
    seconds_elapsed = time_end - time_start
    

    
    return(num_training_obs, num_labels, seconds_elapsed, score)

xs = [0.5,0.3, 0.2, 0.1, 0.09, 0.08, 0.07,0.06, 0.055, 0.05, 0.045]
results = {"range": xs, "obs": [], "places": [], "time": [], "score": []}
for x_range in xs:
    y_range = x_range / 25.0
    a,b,c,d = mlExplore(x_range, y_range)
    results["obs"].append(a)
    results["places"].append(b)
    results["time"].append(round(c,2))
    results["score"].append(round(float(d),2))
pd.DataFrame(results)
    

Expect each model to take about 1.5 seconds
To train 2 Million of them would take 833 hours!
Would need to distribute the problem over a few hundre cores
Would cost £40 to compute on microsoft azure

2e6 * 1.5 /(60*60)

### Tune parameters of Random Forest for the small tree with about 100 observations and 20 places to choose from
 First need to add in the extra variables created from time
